fix: fix failing tests
OpenTofu and Terraform tests are failing with a launch template error. This could be due to numerous reasons, but my assumption is that the AMI does not exist anymore as the previous change was over a year ago. This commit aims to update the AMI to Ubuntu 22.04 in hopes of fixing the failing tests