add max_healthy_percentage - option used during instance refresh (#149)
* add max_healthy_percentage

* fix: fix failing tests

OpenTofu and Terraform tests are failing with a launch template error. This could be due to numerous reasons, but my assumption is that the AMI does not exist anymore as the previous change was over a year ago. This commit aims to update the AMI to Ubuntu 22.04 in hopes of fixing the failing tests

* fix(fix failing tests):   OpenTofu and Terraform tests are failing with a launch template error. This could be due to numerous reasons, but my assumption is that the AMI does not exist anymore as the previous change was over a year ago. This commit aims to update the AMI to Ubuntu 22.04 in hopes of fixing the failing tests

* revert: revert ami changes

---------

Co-authored-by: RoseSecurity <michaeljrosenfeld22@gmail.com>
Co-authored-by: Dan Miller <miller0daniel@gmail.com>